+++ packages/algob/src/types.ts
@@ -687,10 +687,6 @@ export interface SCParams {
   [key: string]: string | bigint
 }
 
-export interface ReplaceParams {
-  [key: string]: string
-}
-
 export interface AnyMap {
   [key: string]: any // eslint-disable-line @typescript-eslint/no-explicit-any
 }