+++ packages/runtime/src/interpreter/interpreter.ts
@@ -9,8 +9,8 @@ import { keyToBytes } from "../lib/parsing";
 import { Stack } from "../lib/stack";
 import { assertMaxCost, parser } from "../parser/parser";
 import {
-  AccountStoreI, ExecutionMode, Operator, SSCAttributesM,
-  StackElem, TEALStack, Txn
+  AccountStoreI, EncTx, ExecutionMode, Operator, SSCAttributesM,
+  StackElem, TEALStack, TxReceipt
 } from "../types";
 import { Op } from "./opcode";
 import { Label } from "./opcode-list";
@@ -41,8 +41,8 @@ export class Interpreter {
   // It is used to provide sub routine functionality
   callStack: Stack<number>;
   labelMap: Map<string, number>; // label string mapped to their respective indexes in instructions array
-  subTxn: Txn | undefined; // "current" inner transaction
-  innerTxns: Txn[]; // executed inner transactions
+  subTxn: EncTx | undefined; // "current" inner transaction
+  innerTxns: EncTx[]; // executed inner transactions
 
   constructor () {
     this.stack = new Stack<StackElem>();
@@ -150,7 +150,7 @@ export class Interpreter {
         account = this.runtime.ctx.state.accounts.get(address);
       } else {
         const accIndex = accountRef - 1n;
-        checkIndexBound(Number(accIndex), this.runtime.ctx.tx.apat as Buffer[], line);
+        checkIndexBound(Number(accIndex), this.runtime.ctx.tx.apat ?? [], line);
         let pkBuffer;
         if (this.runtime.ctx.tx.apat) {
           pkBuffer = this.runtime.ctx.tx.apat[Number(accIndex)];
@@ -433,17 +433,21 @@ export class Interpreter {
     while (this.instructionIndex < this.instructions.length) {
       const instruction = this.instructions[this.instructionIndex];
       instruction.execute(this.stack);
+      const txReceipt = this.runtime.ctx.state.txReceipts.get(this.runtime.ctx.tx.txID) as TxReceipt;
 
       // for teal version >= 4, cost is calculated dynamically at the time of execution
       // for teal version < 4, cost is handled statically during parsing
       dynamicCost += this.lineToCost[instruction.line];
+      if (this.tealVersion < 4) { txReceipt.gas = this.gas; }
       if (this.tealVersion >= 4) {
         if (mode === ExecutionMode.SIGNATURE) {
           assertMaxCost(dynamicCost, mode);
+          txReceipt.gas = dynamicCost;
         } else {
           this.runtime.ctx.pooledApplCost += this.lineToCost[instruction.line];
           const maxPooledApplCost = MaxAppProgramCost * this.runtime.ctx.gtxs.length;
           assertMaxCost(this.runtime.ctx.pooledApplCost, ExecutionMode.APPLICATION, maxPooledApplCost);
+          txReceipt.gas = this.runtime.ctx.pooledApplCost;
         }
       }